simplify get-pages
