--- docs-generator/get-pages.ts
@@ -1,51 +1,0 @@
-/*************************************************************************
- * Copyright 2022 Gravwell, Inc. All rights reserved.
- * Contact: <legal@gravwell.io>
- *
- * This software may be modified and distributed under the terms of the
- * MIT license. See the LICENSE file for details.
- **************************************************************************/
-
-import { getSubfoldersFromFolder } from './folder-manager';
-import { getFilesFromFolder } from './files-manager';
-import { pagesFolderPath, filesFolderPath, readmePathOnDocsGenerator, blackListFolders } from './settings';
-import { Page } from './pages';
-
-/**
- * This function will create a function based on all folders inside './src'
- * @returns all the pages, that will be used inside createTypedocPage()
- */
-export const getPages = (): Array<Page> => {
-	let pages = [];
-	const folders = getSubfoldersFromFolder(filesFolderPath);
-
-	folders.forEach(folderName => {
-		if (!blackListFolders.includes(folderName)) {
-			const folderPath = `${filesFolderPath}/${folderName}`;
-			const files = getFilesFromFolder(folderPath);
-			const subFolders = getSubfoldersFromFolder(folderPath);
-
-			if (files.length > subFolders.length) {
-				pages.push({
-					name: folderName,
-					entryPointStrategy: 'expand',
-					entryPoints: [folderPath],
-					outputDir: `${pagesFolderPath}/${folderName}`,
-					readme: `${readmePathOnDocsGenerator}/${folderName}.md`,
-					customCss: './docs-generator/assets/custom.css',
-				});
-			} else {
-				pages.push({
-					name: folderName,
-					entryPointStrategy: 'resolve',
-					entryPoints: getSubfoldersFromFolder(folderPath).map(subfolder => `${folderPath}/${subfolder}`),
-					outputDir: `${pagesFolderPath}/${folderName}`,
-					readme: `${readmePathOnDocsGenerator}/${folderName}.md`,
-					customCss: './docs-generator/assets/custom.css',
-				});
-			}
-		}
-	});
-
-	return pages;
-};
